requirements.txt
JPype1
install-jdk
requests
loguru